+++ src/main/java/com/robertx22/age_of_exile/player_skills/items/foods/FoodExileEffect.java
@@ -31,12 +31,12 @@ public enum FoodExileEffect {
 
     ELEMENTAL_RESISTANCE("Resistant", "elemental_resist", EffectColor.PURPLE, new OptScaleExactStat(10, new ElementalResist(Elements.Elemental))),
 
-    MAGIC_SHIELD_REGEN("Magicka", "magic_shield_regen", EffectColor.PURPLE, new OptScaleExactStat(20, MagicShieldRegen.getInstance(), ModType.LOCAL_INCREASE)),
-    MANA_REGEN("Arcana", "mana_regen", EffectColor.BLUE, new OptScaleExactStat(20, ManaRegen.getInstance(), ModType.LOCAL_INCREASE)),
-    HEALTH_REGEN("Vitala", "health_regen", EffectColor.RED, new OptScaleExactStat(20, HealthRegen.getInstance(), ModType.LOCAL_INCREASE)),
+    MAGIC_SHIELD_REGEN("Magicka", "magic_shield_regen", EffectColor.PURPLE, new OptScaleExactStat(1, MagicShieldRegen.getInstance()).scale(), new OptScaleExactStat(15, MagicShieldRegen.getInstance(), ModType.LOCAL_INCREASE)),
+    MANA_REGEN("Arcana", "mana_regen", EffectColor.BLUE, new OptScaleExactStat(1, ManaRegen.getInstance()).scale(), new OptScaleExactStat(15, ManaRegen.getInstance(), ModType.LOCAL_INCREASE)),
+    HEALTH_REGEN("Vitala", "health_regen", EffectColor.RED, new OptScaleExactStat(1, HealthRegen.getInstance()).scale(), new OptScaleExactStat(15, HealthRegen.getInstance(), ModType.LOCAL_INCREASE)),
 
     HEALING("Holy", "healing", EffectColor.YELLOW, new OptScaleExactStat(20, HealPower.getInstance())),
-    SPELL_DAMAGE("Enigma", "spell_damage", EffectColor.PURPLE, new OptScaleExactStat(15, SpellDamage.getInstance())),
+    SPELL_DAMAGE("Enigma", "spell_damage", EffectColor.PURPLE, new OptScaleExactStat(10, SpellDamage.getInstance())),
     CRITICAL("Critical", "critical", EffectColor.GREEN, new OptScaleExactStat(5, CriticalHit.getInstance()), new OptScaleExactStat(10, CriticalDamage.getInstance())),
     TREASURE_QUALITY("Treasure", "treasure", EffectColor.YELLOW, new OptScaleExactStat(10, TreasureQuality.getInstance()));